@@ -1,54 +1,30 @@
 {
-  "_from": "content-disposition@0.5.3",
-  "_id": "content-disposition@0.5.3",
-  "_inBundle": false,
-  "_integrity": "sha512-ExO0774ikEObIAEV9kDo50o+79VCUdEB6n6lzKgGwupcVeRlhrj3qGAfwq8G6uBJjkqLrhT0qEYFcWng8z1z0g==",
-  "_location": "/content-disposition",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "content-disposition@0.5.3",
-    "name": "content-disposition",
-    "escapedName": "content-disposition",
-    "rawSpec": "0.5.3",
-    "saveSpec": null,
-    "fetchSpec": "0.5.3"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/content-disposition/-/content-disposition-0.5.3.tgz",
-  "_shasum": "e130caf7e7279087c5616c2007d0485698984fbd",
-  "_spec": "content-disposition@0.5.3",
-  "_where": "C:\\isetr\\ISETR_socialplatform\\back\\node_modules\\express",
-  "author": {
-    "name": "Douglas Christopher Wilson",
-    "email": "doug@somethingdoug.com"
-  },
-  "bugs": {
-    "url": "https://github.com/jshttp/content-disposition/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "safe-buffer": "5.1.2"
-  },
-  "deprecated": false,
+  "name": "content-disposition",
   "description": "Create and parse Content-Disposition header",
+  "version": "0.5.4",
+  "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
+  "license": "MIT",
+  "keywords": [
+    "content-disposition",
+    "http",
+    "rfc6266",
+    "res"
+  ],
+  "repository": "jshttp/content-disposition",
+  "dependencies": {
+    "safe-buffer": "5.2.1"
+  },
   "devDependencies": {
     "deep-equal": "1.0.1",
-    "eslint": "5.10.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.14.0",
-    "eslint-plugin-markdown": "1.0.0-rc.1",
-    "eslint-plugin-node": "7.0.1",
-    "eslint-plugin-promise": "4.0.1",
-    "eslint-plugin-standard": "4.0.0",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "13.0.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.2.0",
+    "eslint-plugin-standard": "4.1.0",
     "istanbul": "0.4.5",
-    "mocha": "5.2.0"
-  },
-  "engines": {
-    "node": ">= 0.6"
+    "mocha": "9.1.3"
   },
   "files": [
     "LICENSE",
@@ -56,24 +32,13 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/jshttp/content-disposition#readme",
-  "keywords": [
-    "content-disposition",
-    "http",
-    "rfc6266",
-    "res"
-  ],
-  "license": "MIT",
-  "name": "content-disposition",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/content-disposition.git"
+  "engines": {
+    "node": ">= 0.6"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "0.5.3"
+    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
+    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/"
+  }
 }
