@@ -1,53 +1,24 @@
 {
-  "_from": "cookie@0.4.0",
-  "_id": "cookie@0.4.0",
-  "_inBundle": false,
-  "_integrity": "sha512-+Hp8fLp57wnUSt0tY0tHEXh4voZRDnoIrZPqlo3DPiI4y9lwg/jqx+1Om94/W6ZaPDOUbnjOt/99w66zk+l1Xg==",
-  "_location": "/cookie",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "cookie@0.4.0",
-    "name": "cookie",
-    "escapedName": "cookie",
-    "rawSpec": "0.4.0",
-    "saveSpec": null,
-    "fetchSpec": "0.4.0"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/cookie/-/cookie-0.4.0.tgz",
-  "_shasum": "beb437e7022b3b6d49019d088665303ebe9c14ba",
-  "_spec": "cookie@0.4.0",
-  "_where": "C:\\isetr\\ISETR_socialplatform\\back\\node_modules\\express",
-  "author": {
-    "name": "Roman Shtylman",
-    "email": "shtylman@gmail.com"
-  },
-  "bugs": {
-    "url": "https://github.com/jshttp/cookie/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "cookie",
   "description": "HTTP server cookie parsing and serialization",
+  "version": "0.4.1",
+  "author": "Roman Shtylman <shtylman@gmail.com>",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>"
+  ],
+  "license": "MIT",
+  "keywords": [
+    "cookie",
+    "cookies"
+  ],
+  "repository": "jshttp/cookie",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "5.16.0",
-    "eslint-plugin-markdown": "1.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "6.1.4"
-  },
-  "engines": {
-    "node": ">= 0.6"
+    "eslint": "6.8.0",
+    "eslint-plugin-markdown": "1.0.2",
+    "mocha": "7.1.1",
+    "nyc": "15.0.1"
   },
   "files": [
     "HISTORY.md",
@@ -55,24 +26,15 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/jshttp/cookie#readme",
-  "keywords": [
-    "cookie",
-    "cookies"
-  ],
-  "license": "MIT",
-  "name": "cookie",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/cookie.git"
+  "engines": {
+    "node": ">= 0.6"
   },
   "scripts": {
     "bench": "node benchmark/index.js",
     "lint": "eslint --plugin markdown --ext js,md .",
-    "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
+    "test": "mocha --reporter spec --bail --check-leaks --ui qunit test/",
+    "test-ci": "nyc --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
     "version": "node scripts/version-history.js && git add HISTORY.md"
-  },
-  "version": "0.4.0"
+  }
 }
