@@ -1,87 +1,69 @@
 {
-  "_from": "qs@6.7.0",
-  "_id": "qs@6.7.0",
-  "_inBundle": false,
-  "_integrity": "sha512-VCdBRNFTX1fyE7Nb6FYoURo/SPe62QCaAyzJvUjwRaIsc+NePBEniHlvxFmmX56+HZphIGtV0XeCirBtpDrTyQ==",
-  "_location": "/qs",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "qs@6.7.0",
     "name": "qs",
-    "escapedName": "qs",
-    "rawSpec": "6.7.0",
-    "saveSpec": null,
-    "fetchSpec": "6.7.0"
-  },
-  "_requiredBy": [
-    "/body-parser",
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/qs/-/qs-6.7.0.tgz",
-  "_shasum": "41dc1a015e3d581f1621776be31afb2876a9b1bc",
-  "_spec": "qs@6.7.0",
-  "_where": "C:\\isetr\\ISETR_socialplatform\\back\\node_modules\\express",
-  "bugs": {
-    "url": "https://github.com/ljharb/qs/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Jordan Harband",
-      "email": "ljharb@gmail.com",
-      "url": "http://ljharb.codes"
+    "description": "A querystring parser that supports nesting and arrays, with a depth limit",
+    "homepage": "https://github.com/ljharb/qs",
+    "version": "6.9.6",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/ljharb/qs.git"
+    },
+    "funding": {
+        "url": "https://github.com/sponsors/ljharb"
+    },
+    "main": "lib/index.js",
+    "contributors": [
+        {
+            "name": "Jordan Harband",
+            "email": "ljharb@gmail.com",
+            "url": "http://ljharb.codes"
+        }
+    ],
+    "keywords": [
+        "querystring",
+        "qs",
+        "query",
+        "url",
+        "parse",
+        "stringify"
+    ],
+    "engines": {
+        "node": ">=0.6"
+    },
+    "dependencies": {},
+    "devDependencies": {
+        "@ljharb/eslint-config": "^17.3.0",
+        "aud": "^1.1.3",
+        "browserify": "^16.5.2",
+        "eclint": "^2.8.1",
+        "eslint": "^7.17.0",
+        "evalmd": "^0.0.19",
+        "for-each": "^0.3.3",
+        "has-symbols": "^1.0.1",
+        "iconv-lite": "^0.5.1",
+        "mkdirp": "^0.5.5",
+        "nyc": "^10.3.2",
+        "object-inspect": "^1.9.0",
+        "qs-iconv": "^1.0.4",
+        "safe-publish-latest": "^1.1.4",
+        "safer-buffer": "^2.1.2",
+        "tape": "^5.1.1"
+    },
+    "scripts": {
+        "prepublish": "safe-publish-latest && npm run dist",
+        "pretest": "npm run --silent readme && npm run --silent lint",
+        "test": "npm run tests-only",
+        "tests-only": "nyc tape 'test/**/*.js'",
+        "posttest": "aud --production",
+        "readme": "evalmd README.md",
+        "postlint": "eclint check * lib/* test/*",
+        "lint": "eslint lib/*.js test/*.js",
+        "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js"
+    },
+    "license": "BSD-3-Clause",
+    "greenkeeper": {
+        "ignore": [
+            "iconv-lite",
+            "mkdirp"
+        ]
     }
-  ],
-  "dependencies": {},
-  "deprecated": false,
-  "description": "A querystring parser that supports nesting and arrays, with a depth limit",
-  "devDependencies": {
-    "@ljharb/eslint-config": "^13.1.1",
-    "browserify": "^16.2.3",
-    "covert": "^1.1.1",
-    "editorconfig-tools": "^0.1.1",
-    "eslint": "^5.15.3",
-    "evalmd": "^0.0.17",
-    "for-each": "^0.3.3",
-    "iconv-lite": "^0.4.24",
-    "mkdirp": "^0.5.1",
-    "object-inspect": "^1.6.0",
-    "qs-iconv": "^1.0.4",
-    "safe-publish-latest": "^1.1.2",
-    "safer-buffer": "^2.1.2",
-    "tape": "^4.10.1"
-  },
-  "engines": {
-    "node": ">=0.6"
-  },
-  "homepage": "https://github.com/ljharb/qs",
-  "keywords": [
-    "querystring",
-    "qs",
-    "query",
-    "url",
-    "parse",
-    "stringify"
-  ],
-  "license": "BSD-3-Clause",
-  "main": "lib/index.js",
-  "name": "qs",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ljharb/qs.git"
-  },
-  "scripts": {
-    "coverage": "covert test",
-    "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js",
-    "lint": "eslint lib/*.js test/*.js",
-    "postlint": "editorconfig-tools check * lib/* test/*",
-    "prepublish": "safe-publish-latest && npm run dist",
-    "pretest": "npm run --silent readme && npm run --silent lint",
-    "readme": "evalmd README.md",
-    "test": "npm run --silent coverage",
-    "tests-only": "node test"
-  },
-  "version": "6.7.0"
 }
