@@ -1,67 +1,33 @@
 {
-  "_from": "raw-body@2.4.0",
-  "_id": "raw-body@2.4.0",
-  "_inBundle": false,
-  "_integrity": "sha512-4Oz8DUIwdvoa5qMJelxipzi/iJIi40O5cGV1wNYp5hvZP8ZN0T+jiNkL0QepXs+EsQ9XJ8ipEDoiH70ySUJP3Q==",
-  "_location": "/raw-body",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "raw-body@2.4.0",
-    "name": "raw-body",
-    "escapedName": "raw-body",
-    "rawSpec": "2.4.0",
-    "saveSpec": null,
-    "fetchSpec": "2.4.0"
-  },
-  "_requiredBy": [
-    "/body-parser"
-  ],
-  "_resolved": "https://registry.npmjs.org/raw-body/-/raw-body-2.4.0.tgz",
-  "_shasum": "a1ce6fb9c9bc356ca52e89256ab59059e13d0332",
-  "_spec": "raw-body@2.4.0",
-  "_where": "C:\\isetr\\ISETR_socialplatform\\back\\node_modules\\body-parser",
-  "author": {
-    "name": "Jonathan Ong",
-    "email": "me@jongleberry.com",
-    "url": "http://jongleberry.com"
-  },
-  "bugs": {
-    "url": "https://github.com/stream-utils/raw-body/issues"
-  },
-  "bundleDependencies": false,
+  "name": "raw-body",
+  "description": "Get and validate the raw body of a readable stream.",
+  "version": "2.4.2",
+  "author": "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)",
   "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Raynos",
-      "email": "raynos2@gmail.com"
-    }
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Raynos <raynos2@gmail.com>"
   ],
+  "license": "MIT",
+  "repository": "stream-utils/raw-body",
   "dependencies": {
-    "bytes": "3.1.0",
-    "http-errors": "1.7.2",
+    "bytes": "3.1.1",
+    "http-errors": "1.8.1",
     "iconv-lite": "0.4.24",
     "unpipe": "1.0.0"
   },
-  "deprecated": false,
-  "description": "Get and validate the raw body of a readable stream.",
   "devDependencies": {
-    "bluebird": "3.5.4",
-    "eslint": "5.16.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.16.0",
-    "eslint-plugin-markdown": "1.0.0",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "6.1.3",
-    "readable-stream": "2.3.6",
-    "safe-buffer": "5.1.2"
+    "bluebird": "3.7.2",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "9.1.3",
+    "nyc": "15.1.0",
+    "readable-stream": "2.3.7",
+    "safe-buffer": "5.2.1"
   },
   "engines": {
     "node": ">= 0.8"
@@ -73,18 +39,10 @@
     "index.d.ts",
     "index.js"
   ],
-  "homepage": "https://github.com/stream-utils/raw-body#readme",
-  "license": "MIT",
-  "name": "raw-body",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/stream-utils/raw-body.git"
-  },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --trace-deprecation --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --trace-deprecation --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --trace-deprecation --reporter spec --check-leaks test/"
-  },
-  "version": "2.4.0"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  }
 }
