@@ -1,49 +1,19 @@
 {
-  "_from": "send@0.17.1",
-  "_id": "send@0.17.1",
-  "_inBundle": false,
-  "_integrity": "sha512-BsVKsiGcQMFwT8UxypobUKyv7irCNRHk1T0G680vk88yf6LBByGcZJOTJCrTP2xVN6yI+XjPJcNuE3V4fT9sAg==",
-  "_location": "/send",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "send@0.17.1",
-    "name": "send",
-    "escapedName": "send",
-    "rawSpec": "0.17.1",
-    "saveSpec": null,
-    "fetchSpec": "0.17.1"
-  },
-  "_requiredBy": [
-    "/express",
-    "/serve-static"
-  ],
-  "_resolved": "https://registry.npmjs.org/send/-/send-0.17.1.tgz",
-  "_shasum": "c1d8b059f7900f7466dd4938bdc44e11ddb376c8",
-  "_spec": "send@0.17.1",
-  "_where": "C:\\isetr\\ISETR_socialplatform\\back\\node_modules\\express",
-  "author": {
-    "name": "TJ Holowaychuk",
-    "email": "tj@vision-media.ca"
-  },
-  "bugs": {
-    "url": "https://github.com/pillarjs/send/issues"
-  },
-  "bundleDependencies": false,
+  "name": "send",
+  "description": "Better streaming static file server with Range and conditional-GET support",
+  "version": "0.17.2",
+  "author": "TJ Holowaychuk <tj@vision-media.ca>",
   "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "James Wyatt Cready",
-      "email": "jcready@gmail.com"
-    },
-    {
-      "name": "Jesús Leganés Combarro",
-      "email": "piranna@gmail.com"
-    }
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "James Wyatt Cready <jcready@gmail.com>",
+    "Jesús Leganés Combarro <piranna@gmail.com>"
+  ],
+  "license": "MIT",
+  "repository": "pillarjs/send",
+  "keywords": [
+    "static",
+    "file",
+    "server"
   ],
   "dependencies": {
     "debug": "2.6.9",
@@ -53,30 +23,25 @@
     "escape-html": "~1.0.3",
     "etag": "~1.8.1",
     "fresh": "0.5.2",
-    "http-errors": "~1.7.2",
+    "http-errors": "1.8.1",
     "mime": "1.6.0",
-    "ms": "2.1.1",
+    "ms": "2.1.3",
     "on-finished": "~2.3.0",
     "range-parser": "~1.2.1",
     "statuses": "~1.5.0"
   },
-  "deprecated": false,
-  "description": "Better streaming static file server with Range and conditional-GET support",
   "devDependencies": {
     "after": "0.8.2",
-    "eslint": "5.16.0",
-    "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.17.2",
-    "eslint-plugin-markdown": "1.0.0",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-promise": "4.1.1",
-    "eslint-plugin-standard": "4.0.0",
-    "istanbul": "0.4.5",
-    "mocha": "6.1.4",
-    "supertest": "4.0.2"
-  },
-  "engines": {
-    "node": ">= 0.8.0"
+    "eslint": "7.32.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.25.3",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.2.0",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "9.1.3",
+    "nyc": "15.1.0",
+    "supertest": "6.1.6"
   },
   "files": [
     "HISTORY.md",
@@ -84,23 +49,13 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/pillarjs/send#readme",
-  "keywords": [
-    "static",
-    "file",
-    "server"
-  ],
-  "license": "MIT",
-  "name": "send",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/pillarjs/send.git"
+  "engines": {
+    "node": ">= 0.8.0"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --check-leaks --reporter spec --bail",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot"
-  },
-  "version": "0.17.1"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  }
 }
